# hld/api/openapi.yaml
openapi: 3.1.0
info:
  title: HumanLayer Daemon REST API
  version: 1.0.0
  description: |
    REST API for HumanLayer daemon operations, providing session management,
    approval workflows, and real-time event streaming capabilities.
  contact:
    name: HumanLayer Support
    url: https://humanlayer.dev
servers:
  - url: http://localhost:7777/api/v1
    description: Local daemon server
paths:
...
  /sessions/{id}:
    get:
      operationId: getSession
      summary: Get session details
      description: Get detailed information about a specific session
      tags:
        - Sessions
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      operationId: updateSession
      summary: Update session settings
      description: |
        Update session settings such as auto-accept mode or archived status.
        Only specified fields will be updated.
      tags:
        - Sessions
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Session updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Invalid status transition or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /anthropic_proxy/{session_id}/v1/messages:
    post:
      summary: Proxy Anthropic API requests for a session
      operationId: proxyAnthropicRequest
      x-manual: true
      tags:
        - proxy-manual
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
          description: Session ID for per-session routing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Proxied response from Anthropic
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
                description: SSE stream for streaming responses
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    sessionId:
      name: id
      in: path
      required: true
      description: Session ID
      schema:
        type: string
      example: sess_abcdef123456
...
  schemas:
...
    Session:
      type: object
      required:
        - id
        - run_id
        - status
        - query
        - created_at
        - last_activity_at
      properties:
        id:
          type: string
          description: Unique session identifier
          example: sess_abcdef123456
        run_id:
          type: string
          description: Unique run identifier
          example: run_xyz789
        claude_session_id:
          type: string
          description: Claude's internal session ID
          example: claude_sess_123
        parent_session_id:
          type: string
          description: Parent session ID if this is a forked session
          example: sess_parent123
        status:
          $ref: '#/components/schemas/SessionStatus'
        query:
          type: string
          description: Initial query that started the session
          example: "Help me refactor this code"
        summary:
          type: string
          description: AI-generated summary of the session
          example: "Refactored authentication module"
        title:
          type: string
          description: User-editable session title
          example: "My Important Task"
        model:
          type: string
          description: Model used for this session
          example: sonnet
        model_id:
          type: string
          description: Full model identifier
          example: claude-opus-4-1-20250805
        working_dir:
          type: string
          description: Working directory for the session
          example: /home/user/project
        additional_directories:
          type: array
          items:
            type: string
          description: Additional directories Claude can access
          example: ["~/.humanlayer/logs", "/var/log/myapp"]
        created_at:
          type: string
          format: date-time
          description: Session creation timestamp
        last_activity_at:
          type: string
          format: date-time
          description: Last activity timestamp
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Session completion timestamp
        error_message:
          type: string
          description: Error message if session failed
        cost_usd:
          type: number
          format: float
          nullable: true
          description: Total cost in USD
          example: 0.05
        input_tokens:
          type: integer
          nullable: true
          description: Number of input tokens
          example: 1000
        output_tokens:
          type: integer
          nullable: true
          description: Number of output tokens
          example: 500
        cache_creation_input_tokens:
          type: integer
          nullable: true
          description: Number of cache creation input tokens
          example: 100
        cache_read_input_tokens:
          type: integer
          nullable: true
          description: Number of cache read input tokens
          example: 50000
        effective_context_tokens:
          type: integer
          nullable: true
          description: Total tokens counting toward context window limit
          example: 51100
        context_limit:
          type: integer
          nullable: true
          description: Context window limit for the model
          example: 168000
        duration_ms:
          type: integer
          nullable: true
          description: Session duration in milliseconds
          example: 45000
        auto_accept_edits:
          type: boolean
          description: Whether edit tools are auto-accepted
          default: false
        dangerously_skip_permissions:
          type: boolean
          description: When true, all tool calls are automatically approved without user consent
          default: false
        dangerously_skip_permissions_expires_at:
          type: string
          format: date-time
          nullable: true
          description: ISO timestamp when dangerously skip permissions mode expires (optional)
        archived:
          type: boolean
          description: Whether session is archived
          default: false
        proxy_enabled:
          type: boolean
          description: Whether proxy is enabled for this session
          default: false
        proxy_base_url:
          type: string
          description: Base URL of the proxy server
          example: https://openrouter.ai/api/v1
        proxy_model_override:
          type: string
          description: Model to use with the proxy
          example: openai/gpt-oss-120b
        editor_state:
          type: string
          nullable: true
          description: JSON blob of editor state for draft sessions
          example: '{"content":"console.log(\"hello\");","cursorPosition":24}'
    SessionStatus:
      type: string
      enum:
        - draft
        - starting
        - running
        - completed
        - failed
        - interrupting
        - interrupted
        - waiting_input
        - discarded
      description: Current status of the session
...
    SessionResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Session'
...
    UpdateSessionRequest:
      type: object
      properties:
        auto_accept_edits:
          type: boolean
          description: Enable/disable auto-accept for edit tools
        dangerously_skip_permissions:
          type: boolean
          description: Enable or disable dangerously skip permissions mode
        dangerously_skip_permissions_timeout_ms:
          type: integer
          format: int64
          nullable: true
          description: Optional timeout in milliseconds for dangerously skip permissions mode
        archived:
          type: boolean
          description: Archive/unarchive the session
        title:
          type: string
          description: Update session title
          example: "Updated Task Name"
        # New fields for model configuration
        model:
          type: string
          description: Model to use (opus, sonnet, or empty for default)
          example: "sonnet"
        model_id:
          type: string
          description: Full model identifier
          example: "claude-3-5-sonnet-20241022"
        # New fields for proxy configuration
        proxy_enabled:
          type: boolean
          description: Enable proxy routing for this session
          example: true
        proxy_base_url:
          type: string
          description: Base URL for proxy service
          example: "https://openrouter.ai/api/v1"
        proxy_model_override:
          type: string
          description: Model identifier for proxy routing
          example: "anthropic/claude-3.5-sonnet"
        proxy_api_key:
          type: string
          description: API key for proxy authentication
          example: "sk-or-..."
        additional_directories:
          type: array
          items:
            type: string
          description: Update additional directories Claude can access
          example: ["~/.humanlayer/logs", "/var/log/myapp"]
        working_dir:
          type: string
          description: Update the working directory for the session
          example: "/Users/john/projects/myapp"
        editor_state:
          type: string
          description: JSON blob of editor state for draft sessions
          example: '{"content":"console.log(\"hello\");","cursorPosition":24}'
        status:
          $ref: '#/components/schemas/SessionStatus'
          description: Update session status (only draft ↔ discarded transitions allowed)
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code (e.g., HLD-101)
          example: HLD-102
        message:
          type: string
          description: Human-readable error message
          example: Session not found
        details:
          type: object
          additionalProperties: true
          description: Additional error context
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: HLD-301
              message: Missing required field 'query'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: HLD-102
              message: Session not found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: HLD-401
              message: Internal server error
              details:
                trace_id: "abc123"
...